refactor: rename CRISPResso2/plotting to CRISPResso2/plots
Matches CRISPRessoPro/plots convention. Updated all imports in
CRISPRessoCORE.py, plots/__init__.py, and test_plot_context.py.
