+++ CRISPResso2/plots/plot_context.py
